+++ src/smc-webapp/client/hub.ts
@@ -10,10 +10,6 @@ declare var Primus: any;
 const { delete_cookie } = require("../misc_page");
 import { QueryParams } from "../misc/query-params";
 import {
-  do_anonymous_setup,
-  should_do_anonymous_setup,
-} from "./anonymous-setup";
-import {
   delete_local_storage,
   from_json_socket,
   to_json_socket,
@@ -23,6 +19,10 @@ import {
 } from "smc-util/misc";
 import { copy_without, uuid } from "smc-util/misc2";
 import * as message from "smc-util/message";
+import {
+  do_anonymous_setup,
+  should_do_anonymous_setup,
+} from "./anonymous-setup";
 
 // Maximum number of outstanding concurrent messages (that have responses)
 // to send at once to hub-websocket.
